Consolidate package structure and document GUI
