+++ tests/test_convert_directory.py
@@ -4,7 +4,7 @@ import subprocess
 import sys
 from pathlib import Path
 
-sys.path.append(str(Path(__file__).resolve().parents[1] / "src"))
+sys.path.append(str(Path(__file__).resolve().parents[1]))
 from video_trim.gui import convert_directory_to_mkv
